fix: app, controllers, errors+status_codes

+++ src/controllers/cards.ts
@@ -1,11 +1,11 @@
 import { Response, Request, NextFunction } from 'express';
+import { STATUS_OK, STATUS_CREATED } from '../utils/status-codes';
 import { RequestCustom } from '../utils/types';
 import Card from '../models/card';
-import User from '../models/user';
-import { WrongDataError, NotFoundError } from '../utils/error';
+import { WrongDataError, NotFoundError, ForbiddenError } from '../utils/error';
 
 export const getCards = (req: Request, res: Response, next: NextFunction) => Card.find({})
-  .then((cards) => { res.send({ cards }); })
+  .then((cards) => { res.status(STATUS_OK).send({ cards }); })
   .catch(next);
 
 export const createCard = (req: Request, res: Response, next: NextFunction) => {
@@ -17,56 +17,49 @@ export const createCard = (req: Request, res: Response, next: NextFunction) => {
       if (!name || !link) {
         throw new WrongDataError('Переданы не все данные');
       }
-      res.send({ card });
+      res.status(STATUS_CREATED).send({ card });
     })
     .catch(next);
 };
 
-export const deleteCardById = (req: Request, res: Response, next: NextFunction) => {
+export const deleteCardById = (req: RequestCustom, res: Response, next: NextFunction) => {
   const id = req.params.cardId;
+  const userId = req.user?._id;
 
   return Card.findById(id)
     .then((card) => {
       if (!card) {
         throw new NotFoundError('Такой карточки не существует');
       }
+      if (String(card?.owner) !== userId) {
+        throw new ForbiddenError('Это не ваша карточка');
+      }
       card?.delete();
-      res.send(card);
+      res.status(STATUS_OK).send(card);
     })
     .catch(next);
 };
 
 export const likeCard = (req: Request, res: Response, next: NextFunction) => {
   const id = (req as RequestCustom)?.user?._id;
-  User.findById(id)
-    .catch(() => {
-      const error = new WrongDataError('Передан неверный пользователь');
-      res.status(error.status).send(error.message);
-    });
 
   return Card.findByIdAndUpdate(req.params.cardId, {
     $addToSet: { likes: id },
   }, {
     new: true,
   })
-    .then((card) => res.send({ card }))
+    .then((card) => res.status(STATUS_CREATED).send({ card }))
     .catch(next);
 };
 
 export const dislikeCard = (req: any, res: Response, next: NextFunction) => {
   const id = req.user._id;
 
-  User.findById(id)
-    .catch(() => {
-      const error = new WrongDataError('Передан неверный пользователь');
-      res.status(error.status).send(error.message);
-    });
-
   return Card.findByIdAndUpdate(req.params.cardId, {
     $pull: { likes: id },
   }, {
     new: true,
   })
-    .then((card) => res.send({ card }))
+    .then((card) => res.status(STATUS_OK).send({ card }))
     .catch(next);
 };
